cli/requirements.txt
click==8.1.7
requests==2.32.5
aws-requests-auth==0.4.3
python-dotenv==1.0.0
rich==13.7.0
PyYAML==6.0.1